Post code review fixes
Override onDowngrade method in DatabaseHelper with full database deletion to avoid crashes when trying to downgrade the app
